Do not remove unmanaged assets
It’s complex to deal with device removal (specially LLDP devices and other scenarios variables, including the NetDiscovery and NetInventory order which then was executed). It would be better to implement a clean up orphans unmanaged devices in GLPI core (it seems that a similar feature was planned for GLPI Inventory plugin as I could see commented on its source code).